+++ angular/lib/src/bootstrap/run.dart
@@ -11,7 +11,7 @@ import '../core/linker.dart'
 import '../core/linker/app_view_utils.dart';
 import '../core/linker/component_resolver.dart' show typeToFactory;
 import '../core/testability/testability.dart';
-import '../core/zone.dart';
+import '../core/zone/ng_zone.dart';
 import '../di/injector/empty.dart';
 import '../di/injector/hierarchical.dart';
 import '../di/injector/injector.dart';
@@ -20,6 +20,9 @@ import '../runtime/dom_events.dart';
 import '../security/dom_sanitization_service.dart';
 
 import 'modules.dart';
+
+/// Used as a "tear-off" of [NgZone].
+NgZone _createNgZone() => NgZone();
 
 /// **INTERNAL ONLY**: Creates a new application-level Injector.
 ///
@@ -33,7 +36,7 @@ import 'modules.dart';
 /// `angular_test` package).
 Injector appInjector(
   InjectorFactory userProvidedInjector, {
-  NgZone Function() createNgZone = createNgZone,
+  NgZone Function() createNgZone = _createNgZone,
 }) {
   // These are the required root services, always provided by AngularDart.
   final Injector minimalInjector = appGlobals.createAppInjector(minimalApp);