+++ angular/lib/src/core/zone.dart
@@ -1,3 +1,0 @@
-// Public API for Zone
-
-export "zone/ng_zone.dart" show NgZone, NgZoneError;