+++ angular/lib/src/core/zone/ng_zone.dart
@@ -2,6 +2,11 @@ import 'dart:async';
 
 import 'package:meta/meta.dart';
 import 'package:stack_trace/stack_trace.dart';
+import 'package:angular/src/facade/exception_handler.dart';
+import 'package:angular/src/runtime.dart';
+
+/// **INTERNAL**: Creates [NgZone] with asynchronous stack traces enabled.
+NgZone debugAsyncStackTracesNgZone() => NgZone._debugAsyncStackTraces();
 
 /// Handles and observes the side-effects of executing callbacks in AngularDart.
 ///
@@ -66,17 +71,39 @@ class NgZone {
   int _pendingMicrotasks = 0;
   final _pendingTimers = <_WrappedTimer>[];
 
-  /// enabled in development mode as they significantly impact perf.
-  NgZone({bool enableLongStackTrace = false}) {
-    _outerZone = Zone.current;
+  factory NgZone({
+    @Deprecated('Remove: Has no effect') bool enableLongStackTrace,
+  }) =>
+      isDevMode && debugAsyncStackTraces
+          ? NgZone._debugAsyncStackTraces()
+          : NgZone._();
 
-    if (enableLongStackTrace) {
-      _innerZone = Chain.capture(() => _createInnerZone(Zone.current),
-          onError: _onErrorWithLongStackTrace);
-    } else {
-      _innerZone = _createInnerZone(Zone.current,
-          handleUncaughtError: _onErrorWithoutLongStackTrace);
+  // TODO(b/134693667): Remove this once a test no longer uses it.
+  NgZone.debugOverrideDoNotUse() {
+    if (!isDevMode) {
+      throw UnsupportedError('Do not use this constructor');
     }
+    _outerZone = Zone.current;
+    _innerZone = _createInnerZone(
+      Zone.current,
+      handleUncaughtError: _onErrorWithoutLongStackTrace,
+    );
+  }
+
+  NgZone._() {
+    _outerZone = Zone.current;
+    _innerZone = _createInnerZone(
+      Zone.current,
+      handleUncaughtError: _onErrorWithoutLongStackTrace,
+    );
+  }
+
+  NgZone._debugAsyncStackTraces() {
+    _outerZone = Zone.current;
+    _innerZone = Chain.capture(
+      () => _createInnerZone(Zone.current),
+      onError: _onErrorWithLongStackTrace,
+    );
   }
 
   /// Whether we are currently executing within this AngularDart zone.